Naively validate json:api root document structure
With a naive validation like this we are likely to capture most of
the cases where a developer passes in an incorrect data structure by
mistake. This does not mean we validate the document, it's more a
convenience feature to provide a better error message in case a document
is malformatted.

The best way to validate a document would be using a json:api schema,
but it has a performance hit and can be done outside the gem scope by a
developer if required.